requirements.txt
asgiref==3.8.1
colorama==0.4.6
coverage==7.8.0
dj-database-url==2.3.0
Django==5.1.7
django-cors-headers==4.7.0
django-filter==25.1
djangorestframework==3.16.0
iniconfig==2.1.0
packaging==24.2
pluggy==1.5.0
psycopg2-binary==2.9.10
pytest==8.3.5
pytest-django==4.11.0
python-dotenv==1.1.0
sqlparse==0.5.3
typing_extensions==4.13.1
tzdata==2025.2
whitenoise==6.9.0
